see desc please: add ton of stuff for betting-pannel
IN GENERAL:
Currently stuff (obv extremely WIP) works in this procedure

1) Start on the Menu scene
2) Choose an option (sel), which leads to the Betting scene
3) Betting scene is useless rn, only displays all nodes representing
upgrades. 2 BilleDNA objects are created and added to dnas[] for each
team. They start with 2 upgrades, those upgrades being rendered as
nodes, using the Nodes {} class. Press space again, and you go to the
fight scene
4) Game scene shows. See PER FILE:./billie.js below for important info

PER FILE:

./bille.js:
add function applied(dna,upgradetree) to add all upgrades from
BilleDNA.upgradetree a beetle's dna as a temp object, hence the deep
copy. The purpose of this is to be able to access the evolved DNA by
procedurally having copies of the basic DNA, and then upgrading those
instead. Very inefficient, but I'm scared of breaking it now.

add an unused getter method, don't touch or uncomment uhh

this._dna is for some reason a property, doesn't have to be. Is just
used to temporarily store the

./index.js
add BILLE_AMOUNT which just contains the amount of biller we want to
fight or whatever

add dnas[], for containing (currently) 2 BilleDNA objects, where the
index corresponds to the respective team - retarded? yes

function Betting(), this is really used as an object (scene) by the
SceneManager library - i have no clue how this works internally, but
it works, so who cares. There's no documentation for SceneManager, but
you can probably clue out how it works per the code here - i did it,
at least. I added comments to this, it should be somewhat clear what
it does.

./node.js
add Node {} class which is just an object describing a node for the
evolution tree in the Betting scene. It should be fairly clear how
this works.